Sourcemaps!
The following plugins had to be replaced to get sourcemaps working:

csso (defunct anyhow) -> clean-css
gulp-wrap -> gulp-wrap-js

Also, responsibility for minification and concatenation has been moved from the build and serve tasks to the asset builders. Done in part because an autoprefixer bug made it necessary, but also because it kinda makes sense.